💄  채팅방 헤더 파일 분리
채팅방 목록, 채팅방 만들기, 채팅방 내부에 각각 사용하는 헤더가 달라서 해당 부분을 분리해줬습니다.
