+++ client/src/components/chat/chat-new-header.tsx
@@ -1,63 +1,13 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable max-len */
-/* eslint-disable no-underscore-dangle */
-/* eslint-disable no-return-assign */
-import { Link, useHistory } from 'react-router-dom';
 import styled, { css } from 'styled-components';
-import { FiMoreHorizontal } from 'react-icons/fi';
-import { BiMessageSquareAdd } from 'react-icons/bi';
 import { useRecoilState, useRecoilValue } from 'recoil';
+import { useHistory } from 'react-router-dom';
 
-import { makeIconToLink } from '@utils/index';
+import ChatHeaderStyle from '@components/chat/style';
 import { postChatRoom } from '@api/index';
 import selectedUserType from '@atoms/chat-selected-users';
 import userType from '@atoms/user';
-
-const ChatHeaderStyle = styled.div`
-  background-color: #B6B6B6;
-
-  width: 100%;
-  height: 80px;
-
-  position: relative;
-
-  p {
-    width: 250px;
-
-    position: absolute;
-    top: 25px;
-    left: 50%;
-    transform: translateX(-50%);
-
-    font-family: 'Nunito';
-    font-weight: Bold;
-    font-size: 32px;
-
-    margin: 0px;
-  }
-`;
-
-const ChatHeaderBtnDiv = styled.div`
-  position: absolute;
-  top: 25px;
-  right: 5%;
-
-  svg{
-    margin: 0px 10px 0px 10px;
-    &:hover {
-      filter: invert(88%) sepia(1%) saturate(121%) hue-rotate(12deg) brightness(62%) contrast(79%);
-    }
-  }
-`;
-
-const chatRoomHeaderBtns = [
-  {
-    Component: BiMessageSquareAdd, link: '/chat-rooms/new', key: 'newChat', size: 32,
-  },
-  {
-    Component: FiMoreHorizontal, link: '/chat-rooms/new', key: 'selector', size: 32,
-  },
-];
 
 const BtnStyle = css`
   position: absolute;
@@ -116,18 +66,7 @@ const CancelBtn = () => {
   return (<CanCelBtnStyle onClick={cancelEvent}>Cancel</CanCelBtnStyle>);
 };
 
-export function ChatRoomHeader() {
-  return (
-    <ChatHeaderStyle>
-      <p>BACK CHANNEL</p>
-      <ChatHeaderBtnDiv>
-        {chatRoomHeaderBtns.map(makeIconToLink)}
-      </ChatHeaderBtnDiv>
-    </ChatHeaderStyle>
-  );
-}
-
-export function NewChatRoomHeader() {
+export default function NewChatRoomHeader() {
   return (
     <ChatHeaderStyle>
       <CancelBtn />
